+++ Changelog.md
@@ -3,6 +3,7 @@
 - New: Items in open Display Cases can now be picked using the "Pick Block" key
 - Fix: Duplication exploit involving keycards
 - Fix: Error involving SecurityCraft's creative tabs when running SecurityCraft alongside certain other mods
+- Fix: Duplication exploit involving Reinforced Pistons
 
 --------------------------Changelog for v1.9.8-beta1 of SecurityCraft--------------------------